+++ test/integration/test_integration_merge.py
@@ -6,7 +6,7 @@ from numpy.testing import assert_almost_equal
 
 from rioxarray import open_rasterio
 from rioxarray.merge import merge_arrays, merge_datasets
-from test.conftest import RASTERIO_GE_125, TEST_INPUT_DATA_DIR
+from test.conftest import RASTERIO_GE_13, RASTERIO_GE_125, TEST_INPUT_DATA_DIR
 
 
 @pytest.mark.parametrize("squeeze", [True, False])
@@ -92,14 +92,16 @@ def test_merge__different_crs(dataset):
         test_sum = merged[merged.rio.vars[0]].sum()
     else:
         test_sum = merged.sum()
-
     if RASTERIO_GE_125:
         assert_almost_equal(
             merged.rio.bounds(),
             (-7300984.0238134, 5003618.5908794, -7224054.1109682, 5050108.6101528),
         )
         assert merged.rio.shape == (84, 139)
-        assert_almost_equal(test_sum, -128605446)
+        if RASTERIO_GE_13:
+            assert_almost_equal(test_sum, -131734881)
+        else:
+            assert_almost_equal(test_sum, -128605446)
     else:
         assert_almost_equal(
             merged.rio.bounds(),